feat: 💄 update missing resource style
