+++ apps/builder/src/page/App/Module/LeftPanel/Components/ToolBar/index.tsx
@@ -3,6 +3,7 @@ import { useSelector } from "react-redux"
 import { getIsILLAGuideMode } from "@/redux/config/configSelector"
 import DebugButton from "../DebugButton"
 import HistoryButton from "../HistoryButton"
+import MissingResourceButton from "../MissingResourceButton"
 import MoreActionButton from "../MoreActionButton"
 import { ToolBarContainerStyle } from "./style"
 
@@ -11,6 +12,7 @@ const ToolBar: FC = () => {
 
   return (
     <div css={ToolBarContainerStyle}>
+      <MissingResourceButton />
       <DebugButton />
       {!isGuideMode && <HistoryButton />}
       {!isGuideMode && <MoreActionButton />}
