+++ apps/builder/src/page/App/components/MissingRosourceModal/interface.ts
@@ -1,0 +1,8 @@
+export interface MissingTipButtonMethod {
+  changeShown: (shown: boolean) => void
+}
+
+export interface IMissingResourceModalProps {
+  shown: boolean
+  changeShown: (shown: boolean) => void
+}
